+++ cloud-sponge/src/main/java/cloud/commandframework/sponge/parser/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Parsers for the Sponge 8 environment.
+ */
+package cloud.commandframework.sponge.parser;
