+++ examples/example-sponge/src/main/java/cloud/commandframework/examples/sponge/CloudExamplePlugin.java
@@ -23,27 +23,16 @@
 //
 package cloud.commandframework.examples.sponge;
 
-import cloud.commandframework.ArgumentDescription;
 import cloud.commandframework.Command;
+import cloud.commandframework.Description;
+import cloud.commandframework.arguments.DefaultValue;
+import cloud.commandframework.arguments.standard.StringParser;
 import cloud.commandframework.context.CommandContext;
 import cloud.commandframework.execution.ExecutionCoordinator;
 import cloud.commandframework.minecraft.extras.MinecraftExceptionHandler;
+import cloud.commandframework.permission.PredicatePermission;
 import cloud.commandframework.sponge.CloudInjectionModule;
 import cloud.commandframework.sponge.SpongeCommandManager;
-import cloud.commandframework.sponge.argument.BlockInputArgument;
-import cloud.commandframework.sponge.argument.BlockPredicateArgument;
-import cloud.commandframework.sponge.argument.DataContainerArgument;
-import cloud.commandframework.sponge.argument.ItemStackPredicateArgument;
-import cloud.commandframework.sponge.argument.MultipleEntitySelectorArgument;
-import cloud.commandframework.sponge.argument.NamedTextColorArgument;
-import cloud.commandframework.sponge.argument.OperatorArgument;
-import cloud.commandframework.sponge.argument.ProtoItemStackArgument;
-import cloud.commandframework.sponge.argument.RegistryEntryArgument;
-import cloud.commandframework.sponge.argument.SinglePlayerSelectorArgument;
-import cloud.commandframework.sponge.argument.UserArgument;
-import cloud.commandframework.sponge.argument.Vector3dArgument;
-import cloud.commandframework.sponge.argument.Vector3iArgument;
-import cloud.commandframework.sponge.argument.WorldArgument;
 import cloud.commandframework.sponge.data.BlockInput;
 import cloud.commandframework.sponge.data.BlockPredicate;
 import cloud.commandframework.sponge.data.ItemStackPredicate;
@@ -94,6 +83,24 @@ import org.spongepowered.math.vector.Vector3d;
 import org.spongepowered.math.vector.Vector3i;
 import org.spongepowered.plugin.builtin.jvm.Plugin;
 
+import static cloud.commandframework.arguments.standard.DoubleParser.doubleParser;
+import static cloud.commandframework.arguments.standard.IntegerParser.integerParser;
+import static cloud.commandframework.arguments.standard.StringParser.greedyStringParser;
+import static cloud.commandframework.arguments.standard.StringParser.stringParser;
+import static cloud.commandframework.sponge.parser.BlockInputParser.blockInputParser;
+import static cloud.commandframework.sponge.parser.BlockPredicateParser.blockPredicateParser;
+import static cloud.commandframework.sponge.parser.DataContainerParser.dataContainerParser;
+import static cloud.commandframework.sponge.parser.ItemStackPredicateParser.itemStackPredicateParser;
+import static cloud.commandframework.sponge.parser.MultipleEntitySelectorParser.multipleEntitySelectorParser;
+import static cloud.commandframework.sponge.parser.NamedTextColorParser.namedTextColorParser;
+import static cloud.commandframework.sponge.parser.OperatorParser.operatorParser;
+import static cloud.commandframework.sponge.parser.ProtoItemStackParser.protoItemStackParser;
+import static cloud.commandframework.sponge.parser.RegistryEntryParser.registryEntryParser;
+import static cloud.commandframework.sponge.parser.SinglePlayerSelectorParser.singlePlayerSelectorParser;
+import static cloud.commandframework.sponge.parser.UserParser.userParser;
+import static cloud.commandframework.sponge.parser.Vector3dParser.vector3dParser;
+import static cloud.commandframework.sponge.parser.Vector3iParser.vector3iParser;
+import static cloud.commandframework.sponge.parser.WorldParser.worldParser;
 import static net.kyori.adventure.text.Component.newline;
 import static net.kyori.adventure.text.Component.space;
 import static net.kyori.adventure.text.Component.text;
@@ -109,11 +116,11 @@ import static net.kyori.adventure.text.format.TextColor.color;
 public final class CloudExamplePlugin {
 
     private static final Component COMMAND_PREFIX = text()
-            .color(color(0x333333))
-            .content("[")
-            .append(text("Cloud-Sponge", color(0xF7CF0D)))
-            .append(text(']'))
-            .build();
+        .color(color(0x333333))
+        .content("[")
+        .append(text("Cloud-Sponge", color(0xF7CF0D)))
+        .append(text(']'))
+        .build();
 
     private final SpongeCommandManager<CommandCause> commandManager;
 
@@ -126,322 +133,318 @@ public final class CloudExamplePlugin {
     public CloudExamplePlugin(final @NonNull Injector injector) {
         // Create child injector with cloud module
         final Injector childInjector = injector.createChildInjector(
-                CloudInjectionModule.createNative(ExecutionCoordinator.simpleCoordinator())
+            CloudInjectionModule.createNative(ExecutionCoordinator.simpleCoordinator())
         );
 
         // Get command manager instance
-        this.commandManager = childInjector.getInstance(Key.get(new TypeLiteral<SpongeCommandManager<CommandCause>>() {
-        }));
+        this.commandManager = childInjector.getInstance(Key.get(new TypeLiteral<SpongeCommandManager<CommandCause>>() {}));
 
         // Use Cloud's enhanced number suggestions
         this.commandManager.parserMapper().cloudNumberSuggestions(true);
 
         // Register minecraft-extras exception handlers
         MinecraftExceptionHandler.create(CommandCause::audience)
-                .defaultHandlers()
-                .decorator(message -> Component.text().append(COMMAND_PREFIX, space(), message).build())
-                .registerTo(this.commandManager);
+            .defaultHandlers()
+            .decorator(message -> Component.text().append(COMMAND_PREFIX, space(), message).build())
+            .registerTo(this.commandManager);
 
         this.registerCommands();
     }
 
     private void registerCommands() {
         this.commandManager.command(this.commandManager.commandBuilder("cloud_test1")
-                .permission("cloud.test1")
-                .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
+            .permission("cloud.test1")
+            .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
         this.commandManager.command(this.commandManager.commandBuilder("cloud_test2")
-                .literal("test")
-                .literal("test1")
-                .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
+            .literal("test")
+            .literal("test1")
+            .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
         final Command.Builder<CommandCause> cloudTest3 = this.commandManager.commandBuilder("cloud_test3");
         final Command.Builder<CommandCause> test = cloudTest3.literal("test");
-        this.commandManager.command(test.argument(StringArgument.single("string_arg"))
-                .literal("test2")
-                .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
+        this.commandManager.command(test.required("string_arg", stringParser())
+            .literal("test2")
+            .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
         this.commandManager.command(test.literal("literal_arg")
-                .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
+            .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
         this.commandManager.command(cloudTest3.literal("another_test")
-                .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
+            .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
         final Command.Builder<CommandCause> cloud = this.commandManager.commandBuilder("cloud");
         this.commandManager.command(cloud.literal("string_test")
-                .argument(StringArgument.single("single"))
-                .argument(StringArgument.quoted("quoted"))
-                .argument(StringArgument.greedy("greedy"))
-                .handler(ctx -> ctx.getSender().audience().sendMessage(text("success"))));
+            .required("single", stringParser(StringParser.StringMode.SINGLE))
+            .required("quoted", stringParser(StringParser.StringMode.QUOTED))
+            .required("greedy", stringParser(StringParser.StringMode.GREEDY))
+            .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
         this.commandManager.command(cloud.literal("int_test")
-                .argument(IntegerArgument.of("any"))
-                .argument(IntegerArgument.<CommandCause>newBuilder("gt0").withMin(1))
-                .argument(IntegerArgument.<CommandCause>newBuilder("lt100").withMax(99))
-                .argument(IntegerArgument.<CommandCause>newBuilder("5to20").withMin(5).withMax(20))
-                .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
+            .required("any", integerParser())
+            .required("gt0", integerParser(1))
+            .required("lt100", integerParser(Integer.MIN_VALUE, 99))
+            .required("5to20", integerParser(5, 20))
+            .handler(ctx -> ctx.sender().audience().sendMessage(text("success"))));
         this.commandManager.command(cloud.literal("enchantment_type_test")
-                .argument(RegistryEntryArgument.of("enchantment_type", EnchantmentType.class, RegistryTypes.ENCHANTMENT_TYPE))
-                .argument(IntegerArgument.optional("level", 1))
-                .handler(ctx -> {
-                    final Object subject = ctx.sender().subject();
-                    if (!(subject instanceof Player)) {
-                        ctx.sender().audience().sendMessage(text("This command is for players only!", RED));
-                        return;
-                    }
-                    final Player player = (Player) subject;
-                    final Hotbar hotbar = player.inventory().hotbar();
-                    final int index = hotbar.selectedSlotIndex();
-                    final Slot slot = hotbar.slot(index).get();
-                    final InventoryTransactionResult.Poll result = slot.poll();
-                    if (result.type() != InventoryTransactionResult.Type.SUCCESS) {
-                        player.sendMessage(text("You must hold an item to enchant!", RED));
-                        return;
-                    }
-                    final ItemStack modified = ItemStack.builder()
-                            .fromItemStack(result.polledItem().createStack())
-                            .add(Keys.APPLIED_ENCHANTMENTS, ImmutableList.of(
-                                    Enchantment.of(
-                                            ctx.<EnchantmentType>get("enchantment_type"),
-                                            ctx.<Integer>get("level")
-                                    )
-                            ))
-                            .build();
-                    slot.set(modified);
-                }));
+            .required("enchantment_type", registryEntryParser(EnchantmentType.class, RegistryTypes.ENCHANTMENT_TYPE))
+            .optional("level", integerParser(), DefaultValue.constant(1))
+            .handler(ctx -> {
+                final Object subject = ctx.sender().subject();
+                if (!(subject instanceof Player)) {
+                    ctx.sender().audience().sendMessage(text("This command is for players only!", RED));
+                    return;
+                }
+                final Player player = (Player) subject;
+                final Hotbar hotbar = player.inventory().hotbar();
+                final int index = hotbar.selectedSlotIndex();
+                final Slot slot = hotbar.slot(index).get();
+                final InventoryTransactionResult.Poll result = slot.poll();
+                if (result.type() != InventoryTransactionResult.Type.SUCCESS) {
+                    player.sendMessage(text("You must hold an item to enchant!", RED));
+                    return;
+                }
+                final ItemStack modified = ItemStack.builder()
+                    .fromItemStack(result.polledItem().createStack())
+                    .add(Keys.APPLIED_ENCHANTMENTS, ImmutableList.of(
+                        Enchantment.of(
+                            ctx.<EnchantmentType>get("enchantment_type"),
+                            ctx.<Integer>get("level")
+                        )
+                    ))
+                    .build();
+                slot.set(modified);
+            }));
         this.commandManager.command(cloud.literal("color_test")
-                .argument(NamedTextColorArgument.of("color"))
-                .argument(StringArgument.greedy("message"))
-                .handler(ctx -> {
-                    ctx.sender().audience().sendMessage(
-                            text(ctx.get("message"), ctx.<NamedTextColor>get("color"))
-                    );
-                }));
+            .required("color", namedTextColorParser())
+            .required("message", greedyStringParser())
+            .handler(ctx -> {
+                ctx.sender().audience().sendMessage(
+                    text(ctx.get("message"), ctx.<NamedTextColor>get("color"))
+                );
+            }));
         this.commandManager.command(cloud.literal("operator_test")
-                .argument(IntegerArgument.of("first"))
-                .argument(OperatorArgument.of("operator"))
-                .argument(IntegerArgument.of("second"))
-                .handler(ctx -> {
-                    final int first = ctx.get("first");
-                    final int second = ctx.get("second");
-                    final Operator operator = ctx.get("operator");
-                    if (!(operator instanceof Operator.Simple)) {
-                        ctx.sender().audience().sendMessage(
-                                text("That type of operator is not applicable here!", RED)
-                        );
-                        return;
-                    }
-                    ctx.sender().audience().sendMessage(text()
-                            .color(AQUA)
-                            .append(text(first))
-                            .append(space())
-                            .append(text(operator.asString(), BLUE))
-                            .append(space())
-                            .append(text(second))
-                            .append(space())
-                            .append(text('→', BLUE))
-                            .append(space())
-                            .append(text(((Operator.Simple) operator).apply(first, second)))
+            .required("first", integerParser())
+            .required("operator", operatorParser())
+            .required("second", integerParser())
+            .handler(ctx -> {
+                final int first = ctx.get("first");
+                final int second = ctx.get("second");
+                final Operator operator = ctx.get("operator");
+                if (!(operator instanceof Operator.Simple)) {
+                    ctx.sender().audience().sendMessage(
+                        text("That type of operator is not applicable here!", RED)
                     );
-                }));
+                    return;
+                }
+                ctx.sender().audience().sendMessage(text()
+                    .color(AQUA)
+                    .append(text(first))
+                    .append(space())
+                    .append(text(operator.asString(), BLUE))
+                    .append(space())
+                    .append(text(second))
+                    .append(space())
+                    .append(text('→', BLUE))
+                    .append(space())
+                    .append(text(((Operator.Simple) operator).apply(first, second)))
+                );
+            }));
         this.commandManager.command(cloud.literal("modifylevel")
-                .argument(OperatorArgument.of("operator"))
-                .argument(DoubleArgument.of("value"))
-                .handler(ctx -> {
-                    final Object subject = ctx.sender().subject();
-                    if (!(subject instanceof Player)) { // todo: a solution to this
-                        ctx.sender().audience().sendMessage(text("This command is for players only!", RED));
-                        return;
-                    }
-                    final Player player = (Player) subject;
-                    final Operator operator = ctx.get("operator");
-                    final double value = ctx.get("value");
-                    if (operator == Operators.ASSIGN.get()) {
-                        player.offer(Keys.EXPERIENCE, (int) value);
-                        return;
-                    }
-                    if (!(operator instanceof Operator.Simple)) {
-                        ctx.sender().audience().sendMessage(
-                                text("That type of operator is not applicable here!", RED)
-                        );
-                        return;
-                    }
-                    final int currentXp = player.get(Keys.EXPERIENCE).get();
-                    player.offer(Keys.EXPERIENCE, (int) ((Operator.Simple) operator).apply(currentXp, value));
-                }));
+            .required("operator", operatorParser())
+            .required("value", doubleParser())
+            .handler(ctx -> {
+                final Object subject = ctx.sender().subject();
+                if (!(subject instanceof Player)) { // todo: a solution to this
+                    ctx.sender().audience().sendMessage(text("This command is for players only!", RED));
+                    return;
+                }
+                final Player player = (Player) subject;
+                final Operator operator = ctx.get("operator");
+                final double value = ctx.get("value");
+                if (operator == Operators.ASSIGN.get()) {
+                    player.offer(Keys.EXPERIENCE, (int) value);
+                    return;
+                }
+                if (!(operator instanceof Operator.Simple)) {
+                    ctx.sender().audience().sendMessage(
+                        text("That type of operator is not applicable here!", RED)
+                    );
+                    return;
+                }
+                final int currentXp = player.get(Keys.EXPERIENCE).get();
+                player.offer(Keys.EXPERIENCE, (int) ((Operator.Simple) operator).apply(currentXp, value));
+            }));
         this.commandManager.command(cloud.literal("selectplayer")
-                .argument(SinglePlayerSelectorArgument.of("player"))
-                .handler(ctx -> {
-                    final Player player = ctx.<SinglePlayerSelector>get("player").getSingle();
-                    ctx.sender().audience().sendMessage(Component.text().append(
-                            text("Display name of selected player: ", GRAY),
-                            player.displayName().get()
-                    ).build());
-                }));
+            .required("player", singlePlayerSelectorParser())
+            .handler(ctx -> {
+                final Player player = ctx.<SinglePlayerSelector>get("player").getSingle();
+                ctx.sender().audience().sendMessage(Component.text().append(
+                    text("Display name of selected player: ", GRAY),
+                    player.displayName().get()
+                ).build());
+            }));
         this.commandManager.command(cloud.literal("world_test")
-                .argument(WorldArgument.of("world"))
-                .handler(ctx -> {
-                    ctx.sender().audience().sendMessage(text(ctx.<ServerWorld>get("world").key().asString()));
-                }));
+            .required("world", worldParser())
+            .handler(ctx -> {
+                ctx.sender().audience().sendMessage(text(ctx.<ServerWorld>get("world").key().asString()));
+            }));
         this.commandManager.command(cloud.literal("test_item")
-                .argument(ProtoItemStackArgument.of("item"))
-                .literal("is")
-                .argument(ItemStackPredicateArgument.of("predicate"))
-                .handler(ctx -> {
-                    final ItemStack item = ctx.<ProtoItemStack>get("item").createItemStack(1, true);
-                    final ItemStackPredicate predicate = ctx.get("predicate");
-                    final Component message = text(builder -> {
-                        builder.append(item.get(Keys.DISPLAY_NAME).orElse(item.type().asComponent()))
-                                .append(space());
-                        if (predicate.test(item)) {
-                            builder.append(text("passes!", GREEN));
-                            return;
-                        }
-                        builder.append(text("does not pass!", RED));
-                    });
-                    ctx.sender().audience().sendMessage(message);
-                }));
-        this.commandManager.command(cloud.literal("test_entity_type")
-                .argument(RegistryEntryArgument.of("type", new TypeToken<EntityType<?>>() {
-                }, RegistryTypes.ENTITY_TYPE))
-                .handler(ctx -> {
-                    ctx.sender().audience().sendMessage(ctx.<EntityType<?>>get("type"));
-                }));
-        final Function<CommandContext<CommandCause>, RegistryHolder> holderFunction = ctx -> ctx.getSender()
-                .location()
-                .map(Location::world)
-                .orElse(Sponge.server().worldManager().defaultWorld());
-        this.commandManager.command(cloud.literal("test_biomes")
-                .argument(RegistryEntryArgument.of("biome", Biome.class, holderFunction, RegistryTypes.BIOME))
-                .handler(ctx -> {
-                    final ResourceKey biomeKey = holderFunction.apply(ctx)
-                            .registry(RegistryTypes.BIOME)
-                            .findValueKey(ctx.get("biome"))
-                            .orElseThrow(IllegalStateException::new);
-                    ctx.sender().audience().sendMessage(text(biomeKey.asString()));
-                }));
-        this.commandManager.command(cloud.literal("test_sounds")
-                .argument(RegistryEntryArgument.of("type", SoundType.class, RegistryTypes.SOUND_TYPE))
-                .handler(ctx -> {
-                    ctx.sender().audience().sendMessage(text(ctx.<SoundType>get("type").key().asString()));
-                }));
-        this.commandManager.command(cloud.literal("summon_villager")
-                .argument(RegistryEntryArgument.of("type", VillagerType.class, RegistryTypes.VILLAGER_TYPE))
-                .argument(RegistryEntryArgument.of("profession", ProfessionType.class, RegistryTypes.PROFESSION_TYPE))
-                .handler(ctx -> {
-                    final ServerLocation loc = ctx.sender().location().orElse(null);
-                    if (loc == null) {
-                        ctx.sender().audience().sendMessage(text("No location!"));
+            .required("item", protoItemStackParser())
+            .literal("is")
+            .required("predicate", itemStackPredicateParser())
+            .handler(ctx -> {
+                final ItemStack item = ctx.<ProtoItemStack>get("item").createItemStack(1, true);
+                final ItemStackPredicate predicate = ctx.get("predicate");
+                final Component message = text(builder -> {
+                    builder.append(item.get(Keys.DISPLAY_NAME).orElse(item.type().asComponent()))
+                        .append(space());
+                    if (predicate.test(item)) {
+                        builder.append(text("passes!", GREEN));
                         return;
                     }
-                    final ServerWorld world = loc.world();
-                    final Villager villager = world.createEntity(EntityTypes.VILLAGER, loc.position());
-                    villager.offer(Keys.VILLAGER_TYPE, ctx.get("type"));
-                    villager.offer(Keys.PROFESSION_TYPE, ctx.get("profession"));
-                    if (world.spawnEntity(villager)) {
-                        ctx.sender().audience().sendMessage(text()
-                                .append(text("Spawned entity!", GREEN))
-                                .append(space())
-                                .append(villager.displayName().get())
-                                .hoverEvent(villager));
-                    } else {
-                        ctx.sender().audience().sendMessage(text("failed to spawn :("));
-                    }
-                }));
+                    builder.append(text("does not pass!", RED));
+                });
+                ctx.sender().audience().sendMessage(message);
+            }));
+        this.commandManager.command(cloud.literal("test_entity_type")
+            .required("type", registryEntryParser(new TypeToken<>() {}, RegistryTypes.ENTITY_TYPE))
+            .handler(ctx -> ctx.sender().audience().sendMessage(ctx.<EntityType<?>>get("type"))));
+        final Function<CommandContext<CommandCause>, RegistryHolder> holderFunction = ctx -> ctx.sender()
+            .location()
+            .map(Location::world)
+            .orElse(Sponge.server().worldManager().defaultWorld());
+        this.commandManager.command(cloud.literal("test_biomes")
+            .required("biome", registryEntryParser(Biome.class, RegistryTypes.BIOME, holderFunction))
+            .handler(ctx -> {
+                final ResourceKey biomeKey = holderFunction.apply(ctx)
+                    .registry(RegistryTypes.BIOME)
+                    .findValueKey(ctx.get("biome"))
+                    .orElseThrow(IllegalStateException::new);
+                ctx.sender().audience().sendMessage(text(biomeKey.asString()));
+            }));
+        this.commandManager.command(cloud.literal("test_sounds")
+            .required("type", registryEntryParser(SoundType.class, RegistryTypes.SOUND_TYPE))
+            .handler(ctx -> {
+                ctx.sender().audience().sendMessage(text(ctx.<SoundType>get("type").key().asString()));
+            }));
+        this.commandManager.command(cloud.literal("summon_villager")
+            .required("type", registryEntryParser(VillagerType.class, RegistryTypes.VILLAGER_TYPE))
+            .required("profession", registryEntryParser(ProfessionType.class, RegistryTypes.PROFESSION_TYPE))
+            .handler(ctx -> {
+                final ServerLocation loc = ctx.sender().location().orElse(null);
+                if (loc == null) {
+                    ctx.sender().audience().sendMessage(text("No location!"));
+                    return;
+                }
+                final ServerWorld world = loc.world();
+                final Villager villager = world.createEntity(EntityTypes.VILLAGER, loc.position());
+                villager.offer(Keys.VILLAGER_TYPE, ctx.get("type"));
+                villager.offer(Keys.PROFESSION_TYPE, ctx.get("profession"));
+                if (world.spawnEntity(villager)) {
+                    ctx.sender().audience().sendMessage(text()
+                        .append(text("Spawned entity!", GREEN))
+                        .append(space())
+                        .append(villager.displayName().get())
+                        .hoverEvent(villager));
+                } else {
+                    ctx.sender().audience().sendMessage(text("failed to spawn :("));
+                }
+            }));
         this.commandManager.command(cloud.literal("vec3d")
-                .argument(Vector3dArgument.of("vec3d"))
-                .handler(ctx -> {
-                    ctx.sender().audience().sendMessage(text(ctx.<Vector3d>get("vec3d").toString()));
-                }));
+            .required("vec3d", vector3dParser())
+            .handler(ctx -> {
+                ctx.sender().audience().sendMessage(text(ctx.<Vector3d>get("vec3d").toString()));
+            }));
         this.commandManager.command(cloud.literal("selectentities")
-                .argument(MultipleEntitySelectorArgument.of("selector"))
-                .handler(ctx -> {
-                    final MultipleEntitySelector selector = ctx.get("selector");
-                    ctx.sender().audience().sendMessage(Component.text().append(
-                            text("Using selector: ", BLUE),
-                            text(selector.inputString()),
-                            newline(),
-                            text("Selected: ", LIGHT_PURPLE),
-                            selector.get().stream()
-                                    .map(e -> e.displayName().get())
-                                    .collect(Component.toComponent(text(", ", GRAY)))
-                    ).build());
-                }));
+            .required("selector", multipleEntitySelectorParser())
+            .handler(ctx -> {
+                final MultipleEntitySelector selector = ctx.get("selector");
+                ctx.sender().audience().sendMessage(Component.text().append(
+                    text("Using selector: ", BLUE),
+                    text(selector.inputString()),
+                    newline(),
+                    text("Selected: ", LIGHT_PURPLE),
+                    selector.get().stream()
+                        .map(e -> e.displayName().get())
+                        .collect(Component.toComponent(text(", ", GRAY)))
+                ).build());
+            }));
 
         this.commandManager.command(cloud.literal("user")
-                .argument(UserArgument.of("user"))
-                .handler(ctx -> {
-                    ctx.sender().audience().sendMessage(text(ctx.<User>get("user").toString()));
-                }));
+            .required("user", userParser())
+            .handler(ctx -> {
+                ctx.sender().audience().sendMessage(text(ctx.<User>get("user").toString()));
+            }));
         this.commandManager.command(cloud.literal("data")
-                .argument(DataContainerArgument.of("data"))
-                .handler(ctx -> {
-                    ctx.sender().audience().sendMessage(text(ctx.<DataContainer>get("data").toString()));
-                }));
+            .required("data", dataContainerParser())
+            .handler(ctx -> {
+                ctx.sender().audience().sendMessage(text(ctx.<DataContainer>get("data").toString()));
+            }));
         this.commandManager.command(cloud.literal("setblock")
-                .permission("cloud.setblock")
-                .argument(Vector3iArgument.of("position"))
-                .argument(BlockInputArgument.of("block"))
-                .handler(ctx -> {
-                    final Vector3i position = ctx.get("position");
-                    final BlockInput input = ctx.get("block");
-                    final Optional<ServerLocation> location = ctx.sender().location();
-                    if (location.isPresent()) {
-                        final ServerWorld world = location.get().world();
-                        input.place(world.location(position));
-                        ctx.sender().audience().sendMessage(text("set block!"));
-                    } else {
-                        ctx.sender().audience().sendMessage(text("no location!"));
-                    }
-                }));
+            .permission("cloud.setblock")
+            .required("position", vector3iParser())
+            .required("block", blockInputParser())
+            .handler(ctx -> {
+                final Vector3i position = ctx.get("position");
+                final BlockInput input = ctx.get("block");
+                final Optional<ServerLocation> location = ctx.sender().location();
+                if (location.isPresent()) {
+                    final ServerWorld world = location.get().world();
+                    input.place(world.location(position));
+                    ctx.sender().audience().sendMessage(text("set block!"));
+                } else {
+                    ctx.sender().audience().sendMessage(text("no location!"));
+                }
+            }));
         this.commandManager.command(cloud.literal("blockinput")
-                .argument(BlockInputArgument.of("block"))
-                .handler(ctx -> {
-                    final BlockInput input = ctx.get("block");
-                    ctx.sender().audience().sendMessage(text(
-                            PaletteTypes.BLOCK_STATE_PALETTE.get().stringifier()
-                                    .apply(RegistryTypes.BLOCK_TYPE.get(), input.blockState())
-                    ));
-                }));
+            .required("block", blockInputParser())
+            .handler(ctx -> {
+                final BlockInput input = ctx.get("block");
+                ctx.sender().audience().sendMessage(text(
+                    PaletteTypes.BLOCK_STATE_PALETTE.get().stringifier()
+                        .apply(RegistryTypes.BLOCK_TYPE.get(), input.blockState())
+                ));
+            }));
         this.commandManager.command(this.commandManager.commandBuilder("gib")
-                .permission("cloud.gib")
-                .argumentPair(
-                        "itemstack",
-                        TypeToken.get(ItemStack.class),
-                        Pair.of("item", "amount"),
-                        Pair.of(ProtoItemStack.class, Integer.class),
-                        (sender, pair) -> {
-                            final ProtoItemStack proto = pair.getFirst();
-                            final int amount = pair.getSecond();
-                            return proto.createItemStack(amount, true);
-                        },
-                        ArgumentDescription.of("The ItemStack to give")
-                )
-                .handler(ctx -> ((Player) ctx.getSender().subject()).inventory().offer(ctx.<ItemStack>get("itemstack"))));
+            .permission("cloud.gib")
+            .requiredArgumentPair(
+                "itemstack",
+                TypeToken.get(ItemStack.class),
+                Pair.of("item", "amount"),
+                Pair.of(ProtoItemStack.class, Integer.class),
+                (sender, pair) -> {
+                    final ProtoItemStack proto = pair.first();
+                    final int amount = pair.second();
+                    return proto.createItemStack(amount, true);
+                },
+                Description.of("The ItemStack to give")
+            )
+            .handler(ctx -> ((Player) ctx.sender().subject()).inventory().offer(ctx.<ItemStack>get("itemstack"))));
         this.commandManager.command(cloud.literal("replace")
-                .permission(cause -> {
-                    // works but error message is ugly
-                    // todo: cause.cause().root() returns DedicatedServer during permission checks?
-                    return cause.subject() instanceof Player;
-                })
-                .argument(BlockPredicateArgument.of("predicate"))
-                .argument(IntegerArgument.of("radius"))
-                .argument(BlockInputArgument.of("replacement"))
-                .handler(ctx -> {
-                    final BlockPredicate predicate = ctx.get("predicate");
-                    final int radius = ctx.get("radius");
-                    final BlockInput replacement = ctx.get("replacement");
+            .permission(PredicatePermission.of(cause -> {
+                // works but error message is ugly
+                // todo: cause.cause().root() returns DedicatedServer during permission checks?
+                return cause.subject() instanceof Player;
+            }))
+            .required("predicate", blockPredicateParser())
+            .required("radius", integerParser())
+            .required("replacement", blockInputParser())
+            .handler(ctx -> {
+                final BlockPredicate predicate = ctx.get("predicate");
+                final int radius = ctx.get("radius");
+                final BlockInput replacement = ctx.get("replacement");
 
-                    // its a player so get is fine
-                    final ServerLocation loc = ctx.sender().location().get();
-                    final ServerWorld world = loc.world();
-                    final Vector3d vec = loc.position();
+                // its a player so get is fine
+                final ServerLocation loc = ctx.sender().location().get();
+                final ServerWorld world = loc.world();
+                final Vector3d vec = loc.position();
 
-                    for (double x = vec.x() - radius; x < vec.x() + radius; x++) {
-                        for (double y = vec.y() - radius; y < vec.y() + radius; y++) {
-                            for (double z = vec.z() - radius; z < vec.z() + radius; z++) {
-                                final ServerLocation location = world.location(x, y, z);
-                                if (predicate.test(location)) {
-                                    location.setBlock(replacement.blockState());
-                                }
+                for (double x = vec.x() - radius; x < vec.x() + radius; x++) {
+                    for (double y = vec.y() - radius; y < vec.y() + radius; y++) {
+                        for (double z = vec.z() - radius; z < vec.z() + radius; z++) {
+                            final ServerLocation location = world.location(x, y, z);
+                            if (predicate.test(location)) {
+                                location.setBlock(replacement.blockState());
                             }
                         }
                     }
-                }));
+                }
+            }));
     }
 
 }
